backend/requirements.txt
fastapi
uvicorn
beautifulsoup4
langchain
langchain-google-genai
sqlalchemy
psycopg2-binary
requests
python-dotenv
pydantic
